refactor(opencode): address message event review feedback
