+++ packages/adapters-opencode-sdk/src/event-stream/message-events/user-display.ts
@@ -2,12 +2,9 @@ import type { AgentUserMessageDisplayPart } from "@openducktor/core";
 import {
   ensureVisibleUserTextDisplayParts,
   mergePreservedAttachmentDisplayParts,
-  normalizeUserMessageDisplayParts,
   readVisibleUserTextFromDisplayParts,
 } from "../../message-normalizers";
 import type { QueuedUserMessageSend, SessionMessageMetadata } from "../../types";
-import type { EventStreamRuntime } from "../shared";
-import { getKnownMessageParts } from "./helpers";
 
 type AttachmentDisplayPart = Extract<AgentUserMessageDisplayPart, { kind: "attachment" }>;
 
@@ -48,33 +45,4 @@ export const buildVisibleUserMessage = (input: {
     displayParts,
     visible: textFromParts.length > 0 ? textFromParts : input.fallbackText,
   };
-};
-
-export const buildKnownUserMessageContent = (
-  runtime: EventStreamRuntime,
-  input: {
-    messageId: string;
-    visible?: string;
-    displayParts?: AgentUserMessageDisplayPart[];
-  },
-): { visible: string; displayParts: AgentUserMessageDisplayPart[] } | null => {
-  const session = runtime.getSession(runtime.externalSessionId);
-  const metadata = session?.messageMetadataById.get(input.messageId);
-  const knownDisplayParts = normalizeUserMessageDisplayParts(
-    getKnownMessageParts(runtime, input.messageId),
-  );
-  const fallbackText = metadata?.text ?? "";
-  const displayParts =
-    input.displayParts ??
-    ensureVisibleUserTextDisplayParts(
-      knownDisplayParts.length > 0 ? knownDisplayParts : (metadata?.displayParts ?? []),
-      fallbackText,
-    );
-  const textFromParts = input.visible ?? readVisibleUserTextFromDisplayParts(displayParts);
-  const visible = textFromParts.length > 0 ? textFromParts : fallbackText;
-  if (visible.trim().length === 0 && displayParts.length === 0) {
-    return null;
-  }
-
-  return { visible, displayParts };
 };
